backend/requirements.txt
# FloraLens — Python Dependencies
# Environment: Python 3.10, CUDA 12.1

# === Core ML ===
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
onnx==1.15.0
onnxruntime==1.16.3

# === API ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# === Data & Image Processing ===
Pillow==10.1.0
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4

# === Experiment Tracking ===
wandb==0.16.1

# === Visualization ===
matplotlib==3.8.2
seaborn==0.13.0

# === Testing ===
pytest==7.4.3
httpx==0.25.2

# === Profiling ===
tensorboard==2.15.1
torch-tb-profiler==0.4.3

# === Utilities ===
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
